week_2_data_ingestion/airflow/requirements.txt
apache-airflow-providers-google
apache-airflow-providers-docker
pyarrow
pandas 
sqlalchemy 
psycopg2-binary